Update file-search to prioritize exact and better fuzzy matches
Fixes #5636

- fixes an issue where better file results were not being displayed
by the file-search due to the limit. The limit meant that if we
ever reached the quota of results allowed some better results
were never to be displayed. Instead, the logic was changed so
that fuzzy matches are sorted by their score (how well they match
the `searchPattern`), then are sent to the front end limited by
the option. This means that all possible exact matches are sent
and the remaining best fuzzy matches along with them until the limit
is reached.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>